+++ geotrek/api/tests/test_v2.py
@@ -9,11 +9,10 @@ from django.contrib.gis.geos import (LineString, MultiLineString, MultiPoint,
                                      Point, Polygon)
 from django.contrib.gis.geos.collections import GeometryCollection
 from django.db import connection
-from django.test import TestCase, RequestFactory
-from django.test.utils import override_settings
+from django.test import TestCase, RequestFactory, override_settings
 from django.urls import reverse
 from django.utils import timezone
-from freezegun.api import freeze_time
+from freezegun import freeze_time
 from mapentity.tests.factories import SuperUserFactory
 from paperclip.models import random_suffix_regexp
 from rest_framework.test import APITestCase, APIClient
@@ -22,8 +21,9 @@ from geotrek import __version__
 from geotrek.api.v2.views.trekking import TrekViewSet
 from geotrek.authent import models as authent_models
 from geotrek.authent.tests import factories as authent_factory
+from geotrek.authent.tests.factories import StructureFactory
 from geotrek.common import models as common_models
-from geotrek.common.tests import factories as common_factory
+from geotrek.common.tests import factories as common_factory, TranslationResetMixin
 from geotrek.common.utils.testdata import (get_dummy_uploaded_document,
                                            get_dummy_uploaded_file,
                                            get_dummy_uploaded_image, get_dummy_uploaded_image_svg)
@@ -37,13 +37,17 @@ from geotrek.infrastructure.tests import factories as infrastructure_factory
 from geotrek.outdoor import models as outdoor_models
 from geotrek.outdoor.tests import factories as outdoor_factory
 from geotrek.sensitivity import models as sensitivity_models
+from geotrek.sensitivity.models import SportPractice
 from geotrek.sensitivity.tests import factories as sensitivity_factory
+from geotrek.sensitivity.tests.factories import SensitiveAreaFactory, MultiPolygonSensitiveAreaFactory, \
+    RegulatorySensitiveAreaFactory
 from geotrek.signage import models as signage_models
 from geotrek.signage.tests import factories as signage_factory
 from geotrek.tourism import models as tourism_models
 from geotrek.tourism.tests import factories as tourism_factory
 from geotrek.trekking import models as trek_models
 from geotrek.trekking.tests import factories as trek_factory
+from geotrek.trekking.tests.base import TrekkingManagerTest
 from geotrek.trekking.tests.factories import PracticeFactory
 from geotrek.zoning import models as zoning_models
 from geotrek.zoning.tests import factories as zoning_factory
@@ -4767,3 +4771,224 @@ class CreateReportsAPITest(TestCase):
         self.assertEqual(report.comment, "We have a problem")
         mock_logger.error.assert_called_with(f"Invalid attachment dummy_file.odt for report {new_report_id} : {{\'attachment_file\': ['File mime type “text/plain” is not allowed for “odt”.']}}")
         self.assertEqual(report.attachments.count(), 0)
+
+
+@freeze_time("2020-01-01")
+class SensitivityAPIv2Test(TranslationResetMixin, TrekkingManagerTest):
+    def setUp(self):
+        super().setUp()
+        self.sensitivearea = SensitiveAreaFactory.create()
+        self.species = self.sensitivearea.species
+        self.pk = self.sensitivearea.pk
+        self.expected_properties = {
+            'create_datetime': self.sensitivearea.date_insert.isoformat().replace('+00:00', 'Z'),
+            'update_datetime': self.sensitivearea.date_update.isoformat().replace('+00:00', 'Z'),
+            'description': "Blabla",
+            "elevation": None,
+            'attachments': [],
+            'contact': '<a href="mailto:toto@tata.com">toto@tata.com</a>',
+            'kml_url': 'http://testserver/api/en/sensitiveareas/{pk}.kml'.format(pk=self.pk),
+            'openair_url': 'http://testserver/api/en/sensitiveareas/{pk}/openair'.format(pk=self.pk),
+            'info_url': self.species.url,
+            'species_id': self.species.id,
+            "name": self.species.name,
+            "period": [False, False, False, False, False, True, True, False, False, False, False, False],
+            'practices': [practice.pk for practice in self.species.practices.all()],
+            'rules': [
+                {'code': 'R1',
+                 'description': None,
+                 'id': self.sensitivearea.rules.all()[0].pk,
+                 'name': 'Rule1',
+                 'pictogram': 'http://testserver/media/picto_rule1.png',
+                 'url': 'http://url.com'},
+                {'code': 'R2',
+                 'description': 'abcdefgh',
+                 'id': self.sensitivearea.rules.all()[1].pk,
+                 'name': 'Rule2',
+                 'pictogram': 'http://testserver/media/picto_rule2.png',
+                 'url': 'http://url.com'}],
+            'provider': '',
+            'structure': 'My structure',
+            'published': True,
+        }
+        self.expected_geom = {
+            'type': 'Polygon',
+            'coordinates': [[
+                [3.0, 46.5],
+                [3.0, 46.500027],
+                [3.0000391, 46.500027],
+                [3.0000391, 46.5],
+                [3.0, 46.5],
+            ]],
+        }
+        self.expected_result = dict(self.expected_properties)
+        self.expected_result['id'] = self.pk
+        self.expected_result['geometry'] = self.expected_geom
+        self.expected_result['url'] = 'http://testserver/api/v2/sensitivearea/{}/?format=json'.format(self.pk)
+        self.expected_geo_result = {
+            'bbox': [3.0, 46.5, 3.0000391, 46.500027],
+            'geometry': self.expected_geom,
+            'type': 'Feature',
+            'id': self.pk,
+            'properties': dict(self.expected_properties),
+        }
+        self.expected_geo_result['properties']['url'] = 'http://testserver/api/v2/sensitivearea/{}/?format=geojson'.format(self.pk)
+
+    @override_settings(SENSITIVITY_OPENAIR_SPORT_PRACTICES=['Practice1', ])
+    def test_detail_sensitivearea(self):
+        url = reverse('apiv2:sensitivearea-detail', args=(self.pk,))
+        params = {'format': 'json', 'period': 'ignore', 'language': 'en'}
+        response = self.client.get(url, params)
+        self.assertJSONEqual(response.content.decode(), self.expected_result)
+
+    @override_settings(SENSITIVITY_OPENAIR_SPORT_PRACTICES=['Practice1', ])
+    def test_detail_sensitivearea_regulatory(self):
+        self.sensitivearea = RegulatorySensitiveAreaFactory.create(species__period01=True)
+        url = reverse('apiv2:sensitivearea-detail', args=(self.sensitivearea.pk,))
+        params = {'format': 'json', 'period': 'ignore', 'language': 'en'}
+        response = self.client.get(url, params)
+        self.assertIsNone(response.json()['species_id'])
+
+    @override_settings(SENSITIVITY_OPENAIR_SPORT_PRACTICES=['Practice1', ])
+    def test_list_sensitivearea(self):
+        url = reverse('apiv2:sensitivearea-list')
+        params = {'format': 'json', 'period': 'ignore', 'language': 'en'}
+        response = self.client.get(url, params)
+        self.assertJSONEqual(response.content.decode(), {
+            'count': 1,
+            'previous': None,
+            'next': None,
+            'results': [self.expected_result],
+        })
+
+    @override_settings(SENSITIVITY_OPENAIR_SPORT_PRACTICES=['Practice1', ])
+    def test_geo_detail_sensitivearea(self):
+        url = reverse('apiv2:sensitivearea-detail', args=(self.pk,))
+        params = {'format': 'geojson', 'period': 'ignore', 'language': 'en'}
+        response = self.client.get(url, params)
+        self.assertJSONEqual(response.content.decode(), self.expected_geo_result)
+
+    @override_settings(SENSITIVITY_OPENAIR_SPORT_PRACTICES=['Practice1', ])
+    def test_geo_list_sensitivearea(self):
+        url = reverse('apiv2:sensitivearea-list')
+        params = {'format': 'geojson', 'period': 'ignore', 'language': 'en'}
+        response = self.client.get(url, params)
+        self.assertJSONEqual(response.content.decode(), {
+            'count': 1,
+            'next': None,
+            'previous': None,
+            'type': 'FeatureCollection',
+            'features': [self.expected_geo_result]
+        })
+
+    def test_no_duplicates_sensitivearea(self):
+        url = reverse('apiv2:sensitivearea-list')
+        params = {'format': 'geojson', 'period': 'ignore', 'language': 'en'}
+        params['practices'] = ','.join([str(p.pk) for p in self.species.practices.all()])
+        response = self.client.get(url, params)
+        self.assertEqual(response.json()['count'], 1, response.json())
+
+    def test_multipolygon(self):
+        sensitivearea = MultiPolygonSensitiveAreaFactory.create()
+        expected_geom = {
+            'type': 'MultiPolygon',
+            'coordinates': [
+                [[
+                    [3.0, 46.5],
+                    [3.0, 46.500027],
+                    [3.0000391, 46.500027],
+                    [3.0000391, 46.5],
+                    [3.0, 46.5],
+                ]],
+                [[
+                    [3.0001304, 46.50009],
+                    [3.0001304, 46.5001171],
+                    [3.0001695, 46.5001171],
+                    [3.0001695, 46.50009],
+                    [3.0001304, 46.50009],
+                ]]
+            ],
+        }
+        url = reverse('apiv2:sensitivearea-detail', args=(sensitivearea.pk,))
+        params = {'format': 'json', 'period': 'ignore', 'language': 'en'}
+        response = self.client.get(url, params)
+        self.assertEqual(response.json()['geometry'], expected_geom)
+
+    @override_settings(SENSITIVITY_OPENAIR_SPORT_PRACTICES=['Practice1', ])
+    def test_list_bubble_sensitivearea(self):
+        url = reverse('apiv2:sensitivearea-list')
+        params = {'format': 'json', 'period': 'ignore', 'language': 'en', 'bubble': 'True'}
+        response = self.client.get(url, params)
+        self.expected_result[u'radius'] = None
+        self.assertJSONEqual(response.content.decode(), {
+            u'count': 1,
+            u'previous': None,
+            u'next': None,
+            u'results': [self.expected_result],
+        })
+
+    def test_list_bubble_sensitivearea_with_point(self):
+        sensitive_area_point = SensitiveAreaFactory.create(geom='SRID=2154;POINT (700040 6600040)',
+                                                           species__period01=True, species__radius=5)
+        url = reverse('apiv2:sensitivearea-list')
+        params = {'format': 'json', 'language': 'en', 'bubble': 'True', 'period': '1'}
+        response = self.client.get(url, params)
+        self.assertEqual(response.json()['count'], 1)
+        self.assertEqual(response.json()['results'][0]['radius'], 5)
+        self.assertEqual(response.json()['results'][0]['name'], sensitive_area_point.species.name)
+
+    def test_list_sportpractice(self):
+        url = reverse('apiv2:sportpractice-list')
+        params = {'format': 'json', 'language': 'en'}
+        response = self.client.get(url, params)
+        sports_practice = SportPractice.objects.all()
+        result_sportpractice = [{'id': sp.id, 'name': sp.name} for sp in sports_practice]
+        self.assertJSONEqual(response.content.decode(), {
+            'count': 2,
+            'previous': None,
+            'next': None,
+            'results': result_sportpractice
+        })
+
+    def test_filters_structure(self):
+        other_structure = StructureFactory.create(name='other')
+        self.sensitivearea_other_structure = SensitiveAreaFactory.create(structure=other_structure)
+        url = reverse('apiv2:sensitivearea-list')
+        params = {'format': 'json', 'period': 'ignore', 'language': 'en'}
+        params['structures'] = other_structure.pk
+        response = self.client.get(url, params)
+        self.assertEqual(response.json()['count'], 1)
+        self.assertEqual(response.json()['results'][0]['name'], self.sensitivearea_other_structure.species.name)
+
+    def test_filters_no_period(self):
+        StructureFactory.create()
+        url = reverse('apiv2:sensitivearea-list')
+        params = {'format': 'json', 'language': 'en'}
+        response = self.client.get(url, params)
+        self.assertEqual(response.json()['count'], 0)
+
+    def test_filters_any_period(self):
+        SensitiveAreaFactory.create()
+        url = reverse('apiv2:sensitivearea-list')
+        params = {'format': 'json', 'period': 'any', 'language': 'en'}
+        response = self.client.get(url, params)
+        self.assertEqual(response.json()['count'], 2)
+
+    def test_filters_specific_period(self):
+        sensitive_area_jf = SensitiveAreaFactory.create(species__period01=True, species__period02=True)
+        SensitiveAreaFactory.create(species__period01=True)
+        SensitiveAreaFactory.create(species__period04=True)
+        url = reverse('apiv2:sensitivearea-list')
+        params = {'format': 'json', 'period': '2,3', 'language': 'en'}
+        response = self.client.get(url, params)
+        self.assertEqual(response.json()['count'], 1)
+        self.assertEqual(response.json()['results'][0]['name'], sensitive_area_jf.species.name)
+
+    def test_filters_no_period_get_month(self):
+        sensitive_area_month = SensitiveAreaFactory.create(**{'species__period01': True})
+        SensitiveAreaFactory.create(**{'species__period02': True})
+        url = reverse('apiv2:sensitivearea-list')
+        params = {'format': 'json', 'language': 'en'}
+        response = self.client.get(url, params)
+        self.assertEqual(response.json()['count'], 1)
+        self.assertEqual(response.json()['results'][0]['name'], sensitive_area_month.species.name)
