+++ geotrek/sensitivity/urls.py
@@ -2,7 +2,6 @@ from django.conf import settings
 from django.urls import path, register_converter
 
 from mapentity.registry import registry
-from rest_framework.routers import DefaultRouter
 
 from geotrek.common.urls import PublishableEntityOptions, LangConverter
 from . import models
@@ -30,9 +29,6 @@ urlpatterns = [
          views.SensitiveAreaOpenAirList.as_view(), name="sensitivearea_openair_list"),
 ]
 
-router = DefaultRouter(trailing_slash=False)
-router.register(r'^api/(?P<lang>[a-z]{2}(-[a-z]{2,4})?)/sensitiveareas', views.SensitiveAreaAPIViewSet, basename='sensitivearea')
-urlpatterns += router.urls
 
 if 'geotrek.trekking' in settings.INSTALLED_APPS:
     urlpatterns.append(path('api/<lang:lang>/treks/<int:pk>/sensitiveareas.geojson',
